KB, use reranker without langchain dependencies (#10758)
